requirements.txt
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.2.25
feedparser==6.0.12
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
magic-filter==1.0.12
multidict==6.7.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
yarl==1.23.0
